pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "brave-search>=0.1.8",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.121.0",
    "google-generativeai>=0.8.5",
    "groq>=0.33.0",
    "langchain>=1.0.3",
    "langchain-groq>=1.0.0",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "openai>=2.7.1",
    "packaging>=25.0",
    "pgvector>=0.4.1",
    "phidata>=2.7.10",
    "psycopg[binary]>=3.2.12",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
